% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{AddClWe}
\alias{AddClWe}
\title{AddClWe}
\usage{
AddClWe(dat)
}
\arguments{
\item{dat}{Imput data frame or matrix.}
}
\value{
Data frame or matrix with two additional columns and rows.
}
\description{
Adding two columns and rows for weights and cluster numbers.
}
\details{
AddClWe
}
\examples{
\dontrun{
X <- matrix( c(0.25,0.1,0.1,0.05,0.2,0.05,0.04,0.03,0.03,0.06,0.02,0.07), 3, 4 )
colnames(X) <- c("j1","j2","j3","j4")
rownames(X) <- c("i1","i2","i3")
X <- AddClWe( X )
}
}
