% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{Cluster}
\alias{Cluster}
\title{Cluster}
\usage{
Cluster(tab, N = 2, byRow = TRUE)
}
\arguments{
\item{N}{Number of clusters.}

\item{byRow}{Choose clustering dimension, rows or columns.}

\item{dat}{Data frame or matrix based on the XGrade() and RemoveClWe() functions.}
}
\value{
Data frame with with additional attributes regarding clustering.
}
\description{
Clustering algorithm.
}
\details{
Cluster
}
\examples{
\dontrun{
X <- matrix( c(0.25,0.1,0.1,0.05,0.2,0.05,0.04,0.03,0.03,0.06,0.02,0.07), 3, 4 )
colnames(X) <- c("j1","j2","j3","j4")
rownames(X) <- c("i1","i2","i3")
X <- AddClWe( X )
XGrade <- TableGCA( X )
Cluster( XGrade, N = 2, byRow = TRUE )
}

}
