% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCA.R
\name{GCA}
\alias{GCA}
\title{GCA}
\usage{
GCA(dat, iter)
}
\arguments{
\item{dat}{Data frame or matrix based on the XGrade() and RemoveClWe() functions.}

\item{iter}{Number of iterations.}
}
\value{
Data frame with columns and rows ordering and Spearman's rho.
}
\description{
GCA algorithm based on Spearman’s rho.
}
\details{
GCA
}
\examples{
\dontrun{
X <- matrix( c(0.25,0.1,0.1,0.05,0.2,0.05,0.04,0.03,0.03,0.06,0.02,0.07), 3, 4 )
colnames(X) <- c("j1","j2","j3","j4")
rownames(X) <- c("i1","i2","i3")
X <- AddClWe( X )
XGrade <- TableGCA( X )
GCA( RemoveClWe( XGrade ), 100 )
}

}
