% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{runShinyGradeStat}
\alias{runShinyGradeStat}
\title{runShinyGradeStat}
\usage{
runShinyGradeStat(launch.browser = F)
}
\arguments{
\item{launch.browser}{By default application is run in a new R window.}
}
\description{
Run Shiny application.
}
\details{
runShinyGradeStat
}
\examples{
\dontrun{
runShinyGradeStat()
}

}
