% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{TableGCA}
\alias{TableGCA}
\title{TableGCA}
\usage{
TableGCA(dat)
}
\arguments{
\item{dat}{Imput data frame or matrix based on the AddClWe() function.}
}
\value{
Data frame or matrix after grade transformation.
}
\description{
Preparing the table after grade transformation.
}
\details{
TableGCA
}
\examples{
\dontrun{
X <- matrix( c(0.25,0.1,0.1,0.05,0.2,0.05,0.04,0.03,0.03,0.06,0.02,0.07), 3, 4 )
colnames(X) <- c("j1","j2","j3","j4")
rownames(X) <- c("i1","i2","i3")
X <- AddClWe( X )
XGrade <- TableGCA( X )
}

}
