% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{TableReorder}
\alias{TableReorder}
\title{TableReorder}
\usage{
TableReorder(dat, rowOrd = NULL, colOrd = NULL)
}
\arguments{
\item{dat}{Data frame or matrix based on the AddClWe() function.}

\item{rowOrd}{New ordering of rows.}

\item{colOrd}{New ordering of columns.}
}
\value{
Data frame or matrix with new ordering of rows and columns.
}
\description{
Reordering of columns and rows based on the GCA() function.
}
\details{
TableReorder
}
\examples{
\dontrun{
X <- matrix( c(0.25,0.1,0.1,0.05,0.2,0.05,0.04,0.03,0.03,0.06,0.02,0.07), 3, 4 )
colnames(X) <- c("j1","j2","j3","j4")
rownames(X) <- c("i1","i2","i3")
X <- AddClWe( X )
XGrade <- TableGCA( X )
ord <- GCA( RemoveClWe( XGrade ), 100 )
XReord1 <- TableReorder( XGrade, ord[1,1:3], ord[1,4:7] )
}

}
